+++ src/lib/hooks/useAddress.ts
@@ -19,9 +19,14 @@ const addressLengthMap: {
     43: "user_address",
     63: "contract_address",
   },
-  /**
-   * @todos Add terra cases
-   */
+  terra2: {
+    44: "user_address",
+    64: "contract_address",
+  },
+  terra2testnet: {
+    44: "user_address",
+    64: "contract_address",
+  },
 };
 
 export const getAddressTypeByLength = (